Separate tests for install_cupy since dependencies such as pytest is not installed at the time of parsing of that file
